# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Consulta de CEP",
        "description": "API para consulta de CEP utilizando o serviço ViaCEP",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor Local"
        }
    ],
    "paths": {
        "/v1/cep/{cep}": {
            "get": {
                "tags": [
                    "CEP"
                ],
                "summary": "Busca informações de um CEP",
                "operationId": "3da9ac64eb3f06c2f43e229c3f6c5784",
                "parameters": [
                    {
                        "name": "cep",
                        "in": "path",
                        "description": "CEP a ser consultado (apenas números)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "pattern": "^[0-9]{8}$"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Informações do CEP encontradas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "cep": {
                                            "type": "string",
                                            "example": "01001-000"
                                        },
                                        "logradouro": {
                                            "type": "string",
                                            "example": "Praça da Sé"
                                        },
                                        "complemento": {
                                            "type": "string",
                                            "example": "lado ímpar"
                                        },
                                        "bairro": {
                                            "type": "string",
                                            "example": "Sé"
                                        },
                                        "localidade": {
                                            "type": "string",
                                            "example": "São Paulo"
                                        },
                                        "uf": {
                                            "type": "string",
                                            "example": "SP"
                                        },
                                        "ibge": {
                                            "type": "string",
                                            "example": "3550308"
                                        },
                                        "gia": {
                                            "type": "string",
                                            "example": "1004"
                                        },
                                        "ddd": {
                                            "type": "string",
                                            "example": "11"
                                        },
                                        "siafi": {
                                            "type": "string",
                                            "example": "7107"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "CEP não encontrado"
                    },
                    "422": {
                        "description": "CEP inválido"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "CEP",
            "description": "CEP"
        }
    ]
}